Fix miniperl.exe.manifest problem on Windows
Commit cb251201d6 inadvertently broke the embedding of the manifest file
in miniperl.exe by changing the target which builds miniperl.exe and hence
the value of $@, which is used in the EMBED_EXE_MANI macro.

It isn't hugely important since the .exe works fine with the .manifest left
alongside it anyway (and miniperl.exe isn't even an installed file either),
but fixing it saves having to .gitignore the .manifest file which was being
left behind rather than embedded and deleted.

We can do this nicely in dmake-speak; unfortunately nmake's version of the
same isn't as nice since it can't handle macros in macro substitutions.