Fatal warnings and errors in destructors
This was brought up in ticket #119311 because I encountered it when
trying to diagnose that bug.

$ ./perl -Ilib -e 'use warnings FATAL=>"all"; DESTROY{die} bless []; exit'
Segmentation fault: 11

Errors that occur in destructors are caught and emitted as warnings.
Fatal warnings turn those warnings into errors.  Those errors are
caught and emitted as warnings.  Fatal warnings turn those warnings
into errors.  Those errors are caught and emitted as warnings.  Fatal
warnings turn those warnings into errors.  Those errors are caught and
emitted as warnings.  Fatal warnings turn those warnings into errors.
Those errors are caught and emitted as warnings.  Have we busted the C
stack yet?

The solution is to remove this warning from any category altogether.
This means it is no longer under the control of ‘use warnings’ and
cannot trigger infinite recursion with fatal warnings.  It is also
arguably more correct for *errors* (which these are) not to be sup-
pressed by default.