Use explicit glob in concise.t.
This was sending a Perl program consisting entirely of '<.>' to
runperl, which on VMS does:

$ perl -e "<.>"
Can't open input file .> as stdin
%RMS-E-FNF, file not found

because the CLI strips the quotes and then the home-grown
redirection code sees the '<' as an invitation to redirect '.>'
to stdin.  That's not readily fixable, so just dodge it here.